\name{mergeDendMergePath}
\alias{mergeDendMergePath}
\title{mergeDendMergePath}
\usage{
  mergeDendMergePath(dend)
}
\arguments{
  \item{dend}{dendrogram}
}
\value{
  the re-merged dendrogram
}
\description{
  re-merge a previously split dendrogram
}
\details{
  This function re-merges a dendrogram that was split using
  the dendMergePath function from this package
}
\author{
  Andreas Leha
}

